fastapi-gateway/requirements.txt
fastapi==0.75.0
uvicorn[standard]==0.17.6
httpx==0.25.1
motor==2.5.1
pydantic==1.9.0
python-dotenv
PyJWT==2.6.0
